Add Group for instruction file
