+++ lib/dustcart.rb
@@ -1,6 +1,7 @@
 require 'dustcart/version'
 require 'dustcart/cli'
 require 'dustcart/dsl'
+require 'dustcart/group'
 require 'dustcart/input'
 
 # Main module
